+++ CHANGELOG.md
@@ -4,7 +4,10 @@
 
 ## 3.0.0RC1
 * Add DEPRECATED.md, EXAMPLES.md, OPTIONS.md, and UPGRADING.md documentation
+* Add filter `tw_examples_html`
+* Add filter `tw_examples`
 * Add filter `tw_get_template_part`
+* Add filter `tw_options_html`
 * Add filter `tw_template_paths`
 * Add filter `tw_widget_categories_args`
 * Add filter `tw_widget_categories_dropdown_args`
@@ -27,12 +30,13 @@
 * Moved email column and data input before URL
 * Put latest Update Notice in readme.txt
 * Readme.txt updates
-* RELATES #103 Example of every option - Add Parsedown prep code
 * Removed customized edit title
 * Removed options: `include_ie7_css`,`use_bxslider`, `disable_animation`, `fade_out_speed`, `fade_in_speed`, `height`, `min_height`, `max_height`
 * Removed pre-2.0.0 migration code
 * Removed shortcodes: `testimonialswidget_list`, `testimonialswidget_widget`
+* Replace `tw_shortcodes` with `tw_examples`
 * Require Aihrus Framework 1.1.5
+* RESOLVE #103 Example of every option
 * RESOLVE #111 bottom_text=false shows false
 * RESOLVE #112 Remove deprecated methods
 * RESOLVE #113 Replace testimonials_widget_ of actions, filters, and functions with tw_